pyproject.toml
[tool.poetry]
name = "jira-sd-discord"
version = "0.1.0"
description = ""
authors = ["Daniel Milnes <thebeanogamer@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.18.1"
Flask = "^1.1.2"
gunicorn = "^20.1.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.4b2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
